Currrent Month Salary Add / View

+++ PFM/include/models/user/usermodal.h
@@ -4,6 +4,6 @@
 #include<mysql.h>
 
 int create_user(MYSQL* conn, MYSQL_STMT* stmt, char user[], char pass[]);
-int authenticate_user(MYSQL* conn, MYSQL_RES* res, MYSQL_ROW row, char user[], char pass[]);
+long int authenticate_user(MYSQL* conn, MYSQL_RES* res, MYSQL_ROW row, char user[], char pass[]);
 
 #endif
